+++ src/main/java/insane96mcp/enhancedai/modules/creeper/TNTLike.java
@@ -1,0 +1,57 @@
+package insane96mcp.enhancedai.modules.creeper;
+
+import insane96mcp.enhancedai.EnhancedAI;
+import insane96mcp.enhancedai.data.EAIData;
+import insane96mcp.enhancedai.modules.Modules;
+import insane96mcp.insanelib.base.Feature;
+import insane96mcp.insanelib.base.LoadFeature;
+import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.DamageTypeTags;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.monster.Creeper;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.event.entity.living.LivingDamageEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+@LoadFeature(module = Modules.Ids.CREEPER, name = "Tnt like creepers", description = "Creepers will ignite when they take explosion damage.")
+public class TNTLike extends Feature {
+	public static final TagKey<EntityType<?>> AFFECTED_ENTITY_TYPES = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("creeper/tnt_like"));
+
+	@Config(min = 0, max = 1d)
+	public static Double chance = 0.25d;
+
+	public static EAIData<Boolean> DATA;
+
+	@Override
+	public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+		super.init(module, enabledByDefault, canBeDisabled);
+		DATA = EAIData.ofBool(this.createDataKey("tnt_like"));
+	}
+
+	@SubscribeEvent(priority = EventPriority.LOWEST)
+	public void eventEntityJoinWorld(EntityJoinLevelEvent event) {
+		if (!this.isEnabled()
+				|| event.getLevel().isClientSide
+				|| !(event.getEntity() instanceof Creeper creeper)
+				|| !creeper.getType().is(AFFECTED_ENTITY_TYPES))
+			return;
+
+		DATA.applyIfAbsent(creeper, creeper.getRandom().nextDouble() < chance);
+	}
+
+	@SubscribeEvent
+	public void livingDamageEvent(LivingDamageEvent event) {
+		if (!this.isEnabled()
+				|| !event.getSource().is(DamageTypeTags.IS_EXPLOSION)
+				|| !(event.getEntity() instanceof Creeper creeper)
+				|| !DATA.get(creeper))
+			return;
+
+		creeper.ignite();
+	}
+
+}